allow showing expired certificates

+++ holder/src/main/java/nl/rijksoverheid/ctr/persistence/database/usecases/RemoveExpiredEventsUseCase.kt
@@ -2,6 +2,7 @@ package nl.rijksoverheid.ctr.persistence.database.usecases
 
 import java.time.Clock
 import java.time.OffsetDateTime
+import nl.rijksoverheid.ctr.holder.usecases.HolderFeatureFlagUseCase
 import nl.rijksoverheid.ctr.persistence.database.HolderDatabase
 import nl.rijksoverheid.ctr.persistence.database.entities.EventGroupEntity
 
@@ -11,10 +12,15 @@ interface RemoveExpiredEventsUseCase {
 
 class RemoveExpiredEventsUseCaseImpl(
     private val clock: Clock,
+    private val featureFlagUseCase: HolderFeatureFlagUseCase,
     private val holderDatabase: HolderDatabase
 ) : RemoveExpiredEventsUseCase {
 
     override suspend fun execute(events: List<EventGroupEntity>) {
+        if (featureFlagUseCase.isInArchiveMode()) {
+            return
+        }
+
         events.forEach {
             if (it.expiryDate != null && it.expiryDate <= OffsetDateTime.now(clock)) {
                 holderDatabase.eventGroupDao().delete(it)
